Add files via upload
An analysis of the coffee capsule and pod market. The analysis is based on tweets from major players in the industry, as well as key features of each brand. The purpose is to figure out what system to create for, and what type of features customers prefer. The focus is on the coffee, rather than the machines. Not looking at financial factors, but market perception.